// src/server/api/methods/bundler.ts
export enum BundlerMethods {
  eth_sendUserOperation = "eth_sendUserOperation",
  eth_estimateUserOperationGas = "eth_estimateUserOperationGas",
  eth_getUserOperationByHash = "eth_getUserOperationByHash",
  eth_getUserOperationReceipt = "eth_getUserOperationReceipt",
  eth_supportedEntryPoints = "eth_supportedEntryPoints",
}