// src/server/api/methods/eth.ts
export enum EthMethods {
  eth_chainId = "eth_chainId",
  eth_getTransactionCount = "eth_getTransactionCount",
  eth_getBalance = "eth_getBalance",
  eth_gasPrice = "eth_gasPrice",
  eth_feeHistory = "eth_feeHistory",
  eth_maxPriorityFeePerGas = "eth_maxPriorityFeePerGas",
  eth_estimateGas = "eth_estimateGas",
  eth_call = "eth_call",
  eth_getTransactionReceipt = "eth_getTransactionReceipt",
  eth_sendRawTransaction = "eth_sendRawTransaction",
}